Impera API request Implementation
Oh god. What have I done?!

I've implemented 3 API requests for the ImperaService. One for logging in, and 2 for games, namely to get a list of games the bot is in, and to look at a specific game.
Might need to delete that last one as it's probably not necessary, but it helped in development.

Remember: always use public variables with Jackson and DTO's

Anyway, After making the API requests, I found I had a lot of DTO's, like a metric tonne of 'em. So I decided to separate every layer in another 3 packages: impera, discord and bot respectively, and keeping with the naming convention set up in the ERD.

I still need to make more packages, but we're getting somewhere.

I also fixed the database not connecting. Credentials are pretty dumb for now, but will change those in prod (obviously). It's localhost anyway.